+++ src/bootstrap.php
@@ -14,16 +14,16 @@
 
 use Origin\Cache\Cache;
 use Origin\Core\Config;
+use Origin\Log\Log;
 
 define('START_TIME', microtime(true));
 
-/*
+
 # This being removed
 error_reporting(E_ALL);
 ini_set('error_reporting', E_ALL);
 ini_set('display_errors', true);
 ini_set('error_log', LOGS);
-*/
 
 /**
  * Load Autoloader
@@ -40,9 +40,14 @@ $errorHandler->register();
 
 require ORIGIN . '/src/Core/functions.php';
 
-/**
- *
- */
+// configure default logging (required)
+// errors during early bootstrap stage can be caught e.g. parse error in application.php
+Log::config('default', [
+    'engine' => 'File',
+    'file' => LOGS . '/application.log'
+]);
+
+// internal caching
 Cache::config('origin', [
     'engine' => 'File',
     'path' => CACHE . '/origin',